Fix: Fixed handling of shout_set_metadata() in case poke request is accepted
This fixes the use of shout_set_metadata() in case TLS Mode is set to
AUTO or AUTO_NO_PLAIN and the server will accept our poke request.

This is normally the case for RoarAudio and XAudioCast and Shoutcast.
While for RoarAudio shout_set_metadata() is not implemented anyway it
is a core requirement for Shoutcast.

Closes: #2298